% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/temporalGraphFunctions.R
\name{addCIs}
\alias{addCIs}
\title{A great function}
\usage{
addCIs(bs, peak = TRUE, opening = TRUE, closing = TRUE, medianFD = TRUE,
  medianLD = TRUE, lwd = 2, code = 3, length = 0.1, ...)
}
\arguments{
\item{bs}{dataframe output from bootstrapFS}

\item{peak}{logical indicating whether to plot confidence interval for peak date.}

\item{opening}{logical indicating whether to plot confidence interval for opening date.}

\item{closing}{logical indicating whether to plot confidence interval for closing date.}

\item{medianFD}{logical indicating whether to plot confidence interval for median first date.}

\item{medianLD}{logical indicating whether to plot confidence interval for median last date.}

\item{lwd}{numeric graphical parameter for width of CI line segment}

\item{code}{numeric graphical parameter for arrows at end of CI line segment}

\item{length}{numeric graphical parameter for length of arrow as end of CI line segment}

\item{...}{Further graphical parameters may also be supplied as arguments.}
}
\description{
This function reads the output of the bootstrapFS function
}
\examples{
\dontrun{addCIs(NULL)}
}
\author{
Stuart Wagenius
}

