% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/temporalDataFunctions.R
\name{bootstrapFS}
\alias{bootstrapFS}
\title{Calculate bootstrap confidence intervals for}
\usage{
bootstrapFS(fs, nboot = 10, bs.ci = c(0.025, 0.975), returnDF = FALSE)
}
\arguments{
\item{fs}{a flowering schedule or 3dpop object.}

\item{nboot}{integer number of times to resample FS.}

\item{bs.ci}{numeric vector of probabilities with values in [0,1].}

\item{returnDF}{logical indicating whether to return dataframe of all resampled parameters.}
}
\description{
A great function
}
\details{
This function reads draws lines in desirable places
}
\examples{
\dontrun{bootstrapFS(NULL)}
}
\author{
Stuart Wagenius
}

