% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{compatibility}
\alias{compatibility}
\title{Mating Type Compatibility of a Population}
\usage{
compatibility(popn, method = "singleLocus", subject = "all",
  averageType = "mean")
}
\arguments{
\item{popn}{a mating scene object}

\item{method}{currently onlu "singleLocus"}

\item{subject}{whether you want pair, individual, population, or all}

\item{averageType}{whether to calculate individual and population proximity
using the mean or median}
}
\description{
Calculate one of several of measures of mating synchrony for a population.
}
\examples{
pop <- simulateScene()
compatibility(pop)
}
\author{
Danny Hanson
}

