% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dailyOMP.R
\name{dailyOMP}
\alias{dailyOMP}
\title{Calculate daily outcrossed mating potential}
\usage{
dailyOMP(scene, k = 4, days = min(scene$start):max(scene$end),
  alpha = 1/13, nn.constant = FALSE)
}
\arguments{
\item{scene}{a matingScene object}

\item{k}{integer, number of nearest neighbors to use in calculating OMP}

\item{days}{the day or range of days to calculate OMP for, default is all days in a scene (see Details)}

\item{alpha}{parameter of exponential decay to be used in calculating OMP (defaults to 1/13)}

\item{nn.constant}{logical; indicates whether the nearest neighbors used in calculations should be the nearest on a given day (\code{nn.constant = FALSE}) or the nearest neighbors over an entire season (\code{nn.constant = TRUE})}
}
\value{
a named matrix with a row for each id and a column for each day, and entries corresponding to ids' OMP each day
}
\description{
dailyOMP generates an OMP object giving the daily outcrossed mating potential of individuals based on k nearest neighbors
}
\details{
The days to calculate OMP for should be input as integers as they are in the start and end columns of a matingScene object. If the number of ids receptive on a day is less than k, OMP will be calculated for the maximum number of neighbors.
}
\examples{
pop <- simulateScene()
omp <- dailyOMP(pop)
omp.1 <- dailyOMP(pop, nn.constant = T) # same neighbors for whole season
}
\author{
Stuart Wagenius
}
\seealso{
\code{\link{makeScene}}
}

