% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporalDataFunctions.R
\name{daysEitherFlowering}
\alias{daysEitherFlowering}
\title{Pairwise total mating timing}
\usage{
daysEitherFlowering(popn, compareToSelf = FALSE)
}
\arguments{
\item{popn}{a 3D population object}

\item{compareToSelf}{whether or not to include self comparisons in the
return value}
}
\value{
a dist-like vector that represents the matrix of
pairwise comparisons
}
\description{
Get all pairwise comparisons of how many days either individual was
flowering
}
\examples{
pop <- generatePop()
daySync <- daysEitherFlowering(pop)
getDistij(daySync, 12, 34)
}
\author{
Danny Hanson
}
\seealso{
\code{\link{getDistij}}
}

