% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/temporalDataFunctions.R
\name{daysSynchronous}
\alias{daysSynchronous}
\title{Pairwise mating timing overlap}
\usage{
daysSynchronous(popn, compareToSelf = FALSE)
}
\arguments{
\item{popn}{a 3D population object}

\item{compareToSelf}{whether or not to include self comparisons in the
return value}
}
\value{
a dist-like vector that represents the matrix of
pairwise comparisons
}
\description{
Get comparisons of mating timing overlap between all pairs
}
\examples{
pop <- generatePop()
daySync <- daysSynchronous(pop)
getDistij(daySync, 12, 34)
}
\author{
Danny Hanson
}
\seealso{
\code{\link{getDistij}}
}

