% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/setUpPopulations.R
\name{generateFS}
\alias{generateFS}
\title{simulate a flowering schedule}
\usage{
generateFS(size = 30, meanSD = "2012-07-12", sdSD = 6, meanDur = 11,
  sdDur = 5, sk = 0)
}
\arguments{
\item{size}{integer number of plants}

\item{meanSD}{date mean start date}

\item{sdSD}{date standard deviation of start date}

\item{meanDur}{numeric duration in days}

\item{sdDur}{standard deviation of duration in days}

\item{sk}{skew of the start date of the population}
}
\value{
data frame with startCol endCol added
}
\description{
generateFS generates a random flowering schedule with standard format
}
\details{
A great function
}
\examples{
generateFS()
\dontrun{generateFS(NULL)}
}
\author{
Stuart Wagenius
}

