% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporalDataFunctions.R
\name{getDaysFlowering}
\alias{getDaysFlowering}
\title{Get flowering days of an individual}
\usage{
getDaysFlowering(popn, ind)
}
\arguments{
\item{popn}{a 3D population object}

\item{ind}{integer. The index/indices of the individual(s) to examine}
}
\value{
a vector containing the start and end days relative to the first
day of flowering of the individual(s)
}
\description{
Get the start and end flowering days of an individual.
}
\examples{
pop <- simulateScene()
getDaysFlowering(pop, 3)
}
\author{
Danny Hanson
}

