% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setUpPopulations.R
\name{getDistElement}
\alias{getDistElement}
\title{Get elements from a dist object}
\usage{
getDistElement(distO, id1, id2)
}
\arguments{
\item{distO}{a dist object or vector made in a dist-like manner}

\item{id1}{character, the id of one of the individuals of interest}

\item{id2}{character, the id of the other individual of interest}
}
\value{
the element that would be in position [i,j] if the dist object
was a matrix
}
\description{
Get the element out of a dist object or dist-like vector that would be
in position [i,j] if the dist object was a matrix Depending on whether
the diagonal is included, the formula to get the [i,j] element is
n*(i-1)-i*(i-1)/2+j if \code{diag} is TRUE or n*(i-1)-i*(i+1)/2+j
if \code{diag} is FALSE.
}
\examples{
pop <- generatePop(size = 200)
daySync <- daysSynchronous(pop)
getDistElement(daySync, "20", "30")
}
\author{
Danny Hanson
}

