% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setUpPopulations.R
\name{getDistij}
\alias{getDistij}
\title{Get elements from a dist object}
\usage{
getDistij(distO, i, j, diag = FALSE)
}
\arguments{
\item{distO}{a dist object or vector made in a dist-like manner}

\item{i}{integer, the row of the object of interest}

\item{j}{integer, the column of the object of interest}

\item{diag}{logical, whether or not the diagonal has been included
in the dist-like vector.}
}
\value{
the element that would be in position [i,j] if the dist object
was a matrix
}
\description{
Get the element out of a dist object or dist-like vector that would be
in position [i,j] if the dist object was a matrix Depending on whether
the diagonal is included, the formula to get the [i,j] element is
n*(i-1)-i*(i-1)/2+j if \code{diag} is FALSE or n*(i-1)-i*(i+1)/2+j
if \code{diag} is FALSE.
}
\examples{
pop <- generatePop(size = 200)
daySync <- daysSynchronous(pop)
getDistij(daySync, 20, 30)

daySyncSelf <- daysSynchronous(pop, TRUE)
getDistij(daySync, 20, 30, TRUE)
}
\author{
Danny Hanson
}

