% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/temporalDataFunctions.R
\name{getDuration}
\alias{getDuration}
\title{duration for each individual}
\usage{
getDuration(popn)
}
\arguments{
\item{popn}{a 3D population object}
}
\value{
a data frame with IDs and durations
}
\description{
Get the duration for each individual in the population
}
\examples{
pop <- generatePop()
getDuration(pop)
}
\author{
Danny Hanson
}

