% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setUpPopulations.R
\name{getijFromInd}
\alias{getijFromInd}
\title{Get i and j from a dist index}
\usage{
getijFromInd(ind, n = -1, distO = NULL, diag = FALSE)
}
\arguments{
\item{ind}{the index within the dist-like vector}

\item{n}{the population size. You need either this or distO}

\item{distO}{a dist object or vector made in a dist-like manner}

\item{diag}{logical, whether or not the diagonal has been included
in the dist-like vector.}
}
\value{
a vector (i, j) of the indeces of the individuals from the original
population
}
\description{
Get i and j given the index within a dist object or dist-like vector
}
\details{
I'm going to admit, this function is highly inefficient because I
couldn't come up with way to invert either n*(i-1)-i*(i-1)/2+j or
n*(i-1)-i*(i+1)/2+j to get a deterministic way to find i and j.
I'm sure something with linear algebra could do it, but it's been
a while, and computers are fast. If it's too slow, I'll write it
in C++
}
\examples{
pop <- generatePop(size = 200)
dayEither <- daysEitherFlowering(pop)
highOnes <- which(dayEither > 10)
getijFromInd(highOnes[1], 200) # only one at a time
getijFromInd(highOnes[1], distO = dayEither) # should be equal
}
\author{
Danny Hanson
}

