% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setUpPopulations.R
\name{getIndex}
\alias{getIndex}
\title{Find indices by id}
\usage{
getIndex(popn, id)
}
\arguments{
\item{popn}{a 3D population object}

\item{id}{character. The id(s) of the individual(s) you want to find}
}
\value{
a integer or vector of integers with the index/indices of
all individuals in the id list
}
\description{
Get the index of an individual with a given id
}
\examples{
pop <- generatePop()
daySync <- daysSynchronous(pop)
getDistij(daySync, 12, 34)
}
\author{
Danny Hanson
}

