% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spatialDataFunctions.R
\name{getPairwiseDistance}
\alias{getPairwiseDistance}
\title{Distance matrix for a population}
\usage{
getPairwiseDistance(popn)
}
\arguments{
\item{popn}{a 3D population object}
}
\value{
a vector representing the lower triangle of the distance matrix
}
\description{
Compute all pairwise distances for a population. This function
is simply a wrapper for \code{dist} that only returns the vector
}
\examples{
pop <- generatePop()
distance <- getPairwiseDistance(pop)
getDistij(distance, 14, 15)
}
\author{
Danny Hanson
}
\seealso{
\code{\link{dist}} \code{\link{getDistij}}
}

