% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialDataFunctions.R
\name{getSpatialDF}
\alias{getSpatialDF}
\title{Get a data frame with only spatial data}
\usage{
getSpatialDF(popn)
}
\arguments{
\item{popn}{a 3d population object}
}
\value{
a data frame with only the id, x, and y columns from the
3d population object
}
\description{
Extract spatial columns from a 3d population object
}
\examples{
pop <- generatePop()
space <- getSpatialDF(pop)
}
\author{
Danny Hanson
}

