% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporalDataFunctions.R
\name{getTemporalDF}
\alias{getTemporalDF}
\title{Get a data frame with only temporal data}
\usage{
getTemporalDF(popn)
}
\arguments{
\item{popn}{a 3d population object}
}
\value{
a data frame with only the id, start, and end columns from the
3d population object
}
\description{
Extract temporal columns from a 3d population object
}
\examples{
pop <- simulateScene()
flowerSched <- getTemporalDF(pop)
}
\author{
Danny Hanson
}

