% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporalDataFunctions.R
\name{individualDailyFlowering}
\alias{individualDailyFlowering}
\title{Flowering by day in a populations}
\usage{
individualDailyFlowering(popn)
}
\arguments{
\item{popn}{a 3D population object}
}
\value{
a matrix where the columns represent all mating days and the rows
represent all individuals in the population. The value for day i and
individual j will be TRUE if that individual was flowering on that day and
is located in position [i,j]
}
\description{
Create a matrix showing which individuals are flowering on a given day
}
\examples{
pop <- simulateScene(size = 10)
individualDailyFlowering(pop)
}
\author{
Danny Hanson
}

