% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setUpPopulations.R
\name{make3d}
\alias{make3d}
\title{update a data frame to a 3dpop}
\usage{
make3d(df, startCol = "startDate", endCol = "endDate", xCol = "x",
  yCol = "y", s1Col = "s1", s2Col = "s2", idcode = "id",
  dateFormat = "\%Y-\%m-\%d")
}
\arguments{
\item{df}{data frame}

\item{startCol}{character name of column with start dates}

\item{endCol}{character name of column with end dates}

\item{xCol}{character name of column with x or E coordinates}

\item{yCol}{character name of column with y or N coordinates}

\item{s1Col}{character name of one column with S-allele}

\item{s2Col}{character name of another column with S-alleles}

\item{idcode}{character name for column with unique identifier}

\item{dateFormat}{character giving either the format of the start and end
dates columns if those columns are characters or the origin if those
columns are numeric (used in as.Date)}
}
\value{
3dpop object, a data frame with columns identified
}
\description{
make3d edits a dataframe into a 3dpop object with standard format
}
\details{
A great function
}
\examples{
\dontrun{makeFS(NULL)}
}
\author{
Stuart Wagenius
}

