% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setUpPopulations.R
\name{makeFS}
\alias{makeFS}
\title{update a data frame to a flowernig schedule}
\usage{
makeFS(df, startCol = "startDate", endCol = "endDate",
  dateFormat = "\%Y-\%m-\%d")
}
\arguments{
\item{df}{data frame}

\item{startCol}{character name of column with start dates}

\item{endCol}{character name of column with end dates}

\item{dateFormat}{character giving either the format of the start and end
dates columns if those columns are characters or the origin if those
columns are numeric (used in as.Date)}
}
\value{
data frame with startCol endCol added
}
\description{
makeFS edits a dataframe into a flowering schedule with standard format
}
\details{
A great function
}
\examples{
flowers <- data.frame(id = c("one", "two"),
                      start = c("2015/07/24", "2015/08/05"),
                      end = c("2015/08/08", "2015/08/11"))
makeFS(flowers, "start", "end", "\%Y/\%m/\%d")
\dontrun{makeFS(NULL)}
}
\author{
Stuart Wagenius
}

