% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setUpPopulations.R
\name{makeScene}
\alias{makeScene}
\title{Create a matingScene object from a data frame}
\usage{
makeScene(df, startCol = "startDate", endCol = "endDate", xCol = "x",
  yCol = "y", s1Col = "s1", s2Col = "s2", idcode = "id",
  dateFormat = "\%Y-\%m-\%d")
}
\arguments{
\item{df}{a data frame containing mating information}

\item{startCol}{character name of column with start dates}

\item{endCol}{character name of column with end dates}

\item{xCol}{character name of column with x or E coordinates}

\item{yCol}{character name of column with y or N coordinates}

\item{s1Col}{character name of one column with S-allele}

\item{s2Col}{character name of another column with S-alleles}

\item{idcode}{character name for column with unique identifier}

\item{dateFormat}{character giving either (1) the format of the start and end
date columns if those columns are characters or (2) the origin for the start
and end date columns if those columns are numeric. It is used in as.Date}
}
\value{
a matingScene object. A data frame with columns identified with
attributes for each of the column names. See details for more information.
}
\description{
Turns a data frame with information about temporal, spatial, or
genetic mating data into a matingScene object using a standard format.
}
\examples{
\dontrun{makeFS(NULL)}
}
\author{
Danny Hanson
}

