% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spatialGraphFunctions.R
\name{mapForDay}
\alias{mapForDay}
\title{Map flowering individuals on a given day}
\usage{
mapForDay(popn, day)
}
\arguments{
\item{popn}{a 3D population object}

\item{day}{integer of which flowering day to plot}
}
\description{
Creates a map that includes all individuals that were flowering on
a given day or set of days
}
\examples{
pop <- generatePop()
mapForDay(pop, 8)
}
\author{
Danny Hanson
}

