% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spatialGraphFunctions.R
\name{mapWithColors}
\alias{mapWithColors}
\title{Map the population with temporal information}
\usage{
mapWithColors(popn, startOrSync = "start")
}
\arguments{
\item{popn}{a 3D population object}

\item{startOrSync}{must be either "start" or "sync". Determines whether
to get colors from start dates or synchrony values}
}
\description{
Create a map of the population where the colors of each point represents
either the start date or the synchrony value of the individual
}
\examples{
pop <- generatePop()
mapWithColors(pop)
}
\author{
Danny Hanson
}

