% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedPotentialPlot.R
\name{mixedPotentialPlot}
\alias{mixedPotentialPlot}
\title{graphical visualization of multiple mating potential objects}
\usage{
mixedPotentialPlot(matPots, showDensity = T, sub.ids = NULL, N = 9,
  sample = "random", sub.labels = FALSE, ind.labels = TRUE,
  lab.cex = 0.5, main = NULL, ...)
}
\value{
nothing

optional arguments for the plot function
}
\description{
Visualize multiple dimensions of mating potential
}
\examples{
pop <- simulateScene()
sync <- synchrony(pop, "augs")
potentialPlot(sync)


}
\author{
Amy Waananen
}

