% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialDataFunctions.R
\name{pairDist}
\alias{pairDist}
\title{Distance Matrix for a Population}
\usage{
pairDist(popn)
}
\arguments{
\item{popn}{a 3D population object}
}
\value{
a matrix of all pairwise comparisons with attribute for order of
ids (idOrder)
}
\description{
Compute all pairwise distances for a population. This function
is simply a wrapper for \code{dist} that only returns a matrix
}
\examples{
pop <- simulateScene()
distance <- pairDist(pop)
}
\author{
Danny Hanson
}
\seealso{
\code{\link{dist}}
}

