% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialGraphFunctions.R
\name{pinpointExtremes}
\alias{pinpointExtremes}
\title{Find earliest and latest flowering}
\usage{
pinpointExtremes(popn)
}
\arguments{
\item{popn}{a 3D population object}
}
\description{
Add lines to show the positions of the earliest and latest flowering
individuals within a population to a map
}
\examples{
pop <- simulateScene()
mapWithColors(pop)
pinpointExtremes(pop)
}
\author{
Danny Hanson
}

