% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/temporalGraphFunctions.R
\name{plotFS}
\alias{plotFS}
\title{Graphical visualization of a flowering schedule}
\usage{
plotFS(fs, opening = NULL, closing = NULL, xlab = "date", ylab = "",
  ...)
}
\arguments{
\item{fs,}{a data frame in flowering schedule format}

\item{opening}{optional date to specify first date of interest}

\item{closing}{optional date to specify last date of interest}

\item{xlab,}{label for the x-axis, defaults to "date"}

\item{ylab,}{label for the y-axis, defaults is blank}

\item{...}{graphical parameters may also be supplied as arguments}
}
\value{
nothing
}
\description{
Create a visualization of the mating schedule of a population
}
\examples{
fs <- generateFS()
plotFS(fs)
\dontrun{plotFS(NULL)}
}
\author{
Stuart Wagenius
}

