% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialGraphFunctions.R
\name{plotMap}
\alias{plotMap}
\title{graphical visualization of a spatial disribution}
\usage{
plotMap(pp, xlab = "E", ylab = "N", pch = NULL, ...)
}
\arguments{
\item{pp}{dataframe in 3dpop format, at least with x and y columns}

\item{xlab}{character label for the x-axis, defaults to "E"}

\item{ylab}{character label for the y-axis, defaults to "N"}

\item{pch}{optional plotting character}

\item{...}{graphical parameters may also be supplied as arguments}
}
\value{
nothing

optional arguments for the plot function
}
\description{
plotMap makes a map of a population
}
\examples{
pop <- simulateScene()
plotMap(pop)
\dontrun{plotMap(NULL)}
}
\author{
Stuart Wagenius
}
\seealso{
see generic function \code{\link{points}} for values of \code{pch}
}

