% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/temporalGraphFunctions.R
\name{pointsFS}
\alias{pointsFS}
\title{add daily summary of flowering schedule to a plot}
\usage{
pointsFS(fs, peak = FALSE, ...)
}
\arguments{
\item{fs}{a list in flowering schedule or 3dpop format}

\item{peak}{logical add vertical line at peak date, default is false}

\item{...}{graphical parameters may also be supplied as arguments}
}
\value{
the output of summaryFS invisibly
}
\description{
pointsFS adds points to a flowering schedule figure
}
\details{
A great function. It invisibly returns the same thing as summaryFS.
}
\examples{
pop <- generatePop()
plotFS(pop)
pointsFS(pop, type = "b")
}
\author{
Stuart Wagenius
}
\seealso{
see functions \code{\link{plotFS}} and \code{\link{summaryFS}}
}

