% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potentialPlot.R
\name{potentialPlot}
\alias{potentialPlot}
\title{graphical visualization of a mating potential object}
\usage{
potentialPlot(matPot, subject = c("pair", "ind"), plotType = "auto",
  showDensity = T, sub.ids = NULL, N = 9, sample = "random",
  ind.labels = TRUE, lab.cex = 0.5, main = NULL, ...)
}
\arguments{
\item{matPot}{a mating potential object}

\item{subject}{a character string indicating whether the subject to be visualized is individuals or all pairwise interactions}

\item{plotType}{a character string indivating the plots to be displayed. Options are histogram ('hist'), network diagram ('net')}

\item{showDensity}{logical. If true (default), plots probability density over histogram.}

\item{sub.ids}{a vector containing the ids of individuals to be represented in pairwise potential plots}

\item{N}{a positive number, the number of individuals to sample if sub.ids = 'random'}

\item{sample}{a character string specifying how to choose a subset of individuals to be represented in pairwise potential plots. Possible values are "random" (default) or "all".}

\item{lab.cex}{parameter indicating label size relative to plot}

\item{main}{the main title (on top of plot)}

\item{...}{optional arguments for the plot function}
}
\value{
nothing
}
\description{
Visualize mating potential
}
\examples{
pop <- simulateScene()
sync <- synchrony(pop, "augs")
potentialPlot(sync)


}
\author{
Amy Waananen
}
\seealso{
see generic function \code{\link{points}} for values of \code{pch}
}

