% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/temporalDataFunctions.R
\name{summaryFS}
\alias{summaryFS}
\title{summary of a flowering schedule}
\usage{
summaryFS(fs, opening = NULL, closing = NULL)
}
\arguments{
\item{fs,}{a data frame in flowering schedule format}

\item{opening}{optional date to specify first date of interest}

\item{closing}{optional date to specify last date of interest}
}
\value{
a names list with info about the flowering schedule
}
\description{
summaryFS returns standard info about a flowering schedule
}
\details{
A great function
}
\examples{
fs <- generateFS()
summaryFS(fs)
\dontrun{summaryFS(NULL)}
}
\author{
Stuart Wagenius
}

