% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/temporalGraphFunctions.R
\name{vlinesFS}
\alias{vlinesFS}
\title{add vertical lines to a flowering schedule plot}
\usage{
vlinesFS(fs, peak = TRUE, medianFD = TRUE, medianLD = TRUE, ...)
}
\arguments{
\item{fs}{a flowering schedule or 3dpop object}

\item{peak}{logical indicating whether to draw line at the peak date.}

\item{medianFD}{logical indicating whether to draw a line at the median first date.}

\item{medianLD}{logical indicating whether to draw a line at the median last date.}

\item{...}{Further graphical parameters may also be supplied as arguments.}
}
\description{
A great function
}
\details{
This function reads draws lines in desirable places
}
\examples{
\dontrun{vlinesFS(NULL)}
}
\author{
Stuart Wagenius
}

