% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setUpPopulations.R
\name{makeScene}
\alias{makeScene}
\title{Create a matingScene object from a data frame}
\usage{
makeScene(df, startCol = "start", endCol = "end", xCol = "x",
  yCol = "y", s1Col = "s1", s2Col = "s2", idCol = "id",
  dateFormat = "\%Y-\%m-\%d")
}
\arguments{
\item{df}{a data frame containing mating information}

\item{startCol}{character name of column with start dates}

\item{endCol}{character name of column with end dates}

\item{xCol}{character name of column with x or E coordinates}

\item{yCol}{character name of column with y or N coordinates}

\item{s1Col}{character name of one column with S-allele}

\item{s2Col}{character name of another column with S-alleles}

\item{idCol}{character name for column with unique identifier}

\item{dateFormat}{character giving either (1) the format of the start and end
date columns if those columns are characters or (2) the origin for the start
and end date columns if those columns are numeric. It is used in as.Date}
}
\value{
a matingScene data frame. A data frame with column names in a
specific format. There are also attributes giving what type of mating
information is in the data frame, what the original column names were,
and the origin of the date columns. See details for more information.
}
\description{
Turns a data frame with information about temporal, spatial, or
genetic mating data into a matingScene object using a standard format.
}
\details{
The columns are named id, start, end, x, y, s1, and s2 for
idCol, startCol, endCol, xCol, yCol, s1Col, and s2Col respectively.
The attributes "t", "s", and "c" will be set to TRUE if the data frame
has temporal, spatial, or genetic compatibility data, respectively and
will be FALSE otherwise. The attribute originalNames contains all the
names of the original data frame.
The start and end columns will be changed to integers relative to the start
day of the population. So the first day of the first individual to become
receptive will be one and so on. The attribute origin contains the
origin that can be used when converting the columns start and end
from integers to dates.
}
\examples{
\dontrun{makeFS(NULL)}
}
\author{
Danny Hanson
}

