% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporalDataFunctions.R
\name{overlap}
\alias{overlap}
\title{Pairwise mating timing comparison}
\usage{
overlap(popn, overlapOrTotal = c("overlap", "total"), compareToSelf = FALSE)
}
\arguments{
\item{popn}{a 3D population object}

\item{overlapOrTotal}{whether to calculate the number of days that each
pair was overlapping in mating receptivity or the total number of days
that either individual was receptive}

\item{compareToSelf}{whether or not to include self comparisons in the
return value}
}
\value{
a matrix containing all pairwise comparisons. If compareToSelf
is FALSE then there will be n rows and n-1 columns and indexing
result[i,j] where j > i must be done with result[i,j-1], whether result
is the return value of overlap. There is one attribute "idOrder" which
gives the order of the id column in popn at the time of the function call.
This can be useful to find certain elements in the matrix, see examples.
}
\description{
Get comparisons of mating timing between all pairs
}
\examples{
pop <- generatePop()
pop <- pop[order(pop$start),]
daysSync <- overlap(pop)
indices <- which(attr(daysSync, "inOrder") \%in\% c(1, 4))
if (indices[1] <= indices[2]) {
  daysSync[indices[1], indices[2]]
} else {
  daysSync[indices[1], indices[2]-1]
}
}
\author{
Danny Hanson
}
\seealso{
\code{\link{getDistij}}
}

