% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialDataFunctions.R
\name{pairDist}
\alias{pairDist}
\title{Distance Matrix for a mating scene}
\usage{
pairDist(scene)
}
\arguments{
\item{scene}{a matingScene object}
}
\value{
a matrix of all pairwise comparisons with attributes for order of
identifiers (idOrder)
}
\description{
Compute all pairwise distances for a population. This function
is simply a wrapper for \code{dist} that returns only a matrix
}
\examples{
pop <- simulateScene()
distance <- pairDist(pop)
}
\seealso{
\code{\link{dist}}
}

