% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialDataFunctions.R
\name{proximity}
\alias{proximity}
\title{Spatial Proximity of a Population}
\usage{
proximity(popn, method, proximityFun = NULL, averageType = "mean",
  subject = "all")
}
\arguments{
\item{popn}{a 3D population object}

\item{method}{currently one of "maxProp", and "maxPropSqrd"}

\item{proximityFun}{a function used to calculate proximity}

\item{averageType}{whether to calculate individual and population proximity
using the mean or median}

\item{subject}{whether you want pair, individual, population, or all}
}
\description{
Calculate one of several of measures of mating synchrony for a population.
}
\examples{
pop <- simulateScene()
proximity(pop)
}
\author{
Danny Hanson
}

