% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialDataFunctions.R
\name{proximity}
\alias{proximity}
\title{Spatial proximity of a population}
\usage{
proximity(popn, proximityFun = NULL, proximityType = c("individual",
  "population", "both"))
}
\arguments{
\item{popn}{a 3D population object}

\item{proximityFun}{a function used to calculate proximity}

\item{proximityType}{whether you want individual, population, or both}
}
\description{
Currently the function doesn't do anything. If you like returning null,
I would use it a lot.
}
\examples{
pop <- generatePop()
proximity(pop)
}
\author{
No Body
}

